+++ artemis-matrix/src/test/java/com/artemis/component/ExtPosition.java
@@ -1,6 +1,0 @@
-package com.artemis.component;
-
-import com.artemis.Component;
-
-public class ExtPosition extends Position {
-}