+++ artemis-matrix/src/test/java/com/artemis/component/Velocity.java
@@ -1,6 +1,0 @@
-package com.artemis.component;
-
-import com.artemis.Component;
-
-public class Velocity extends Component {
-}