+++ artemis-matrix/src/test/java/com/artemis/manager/SomeManager.java
@@ -1,6 +1,0 @@
-package com.artemis.manager;
-
-import com.artemis.Manager;
-
-public class SomeManager extends Manager {
-}