+++ artemis-weaver/src/main/java/com/artemis/meta/ClassMetadata.java
@@ -47,6 +47,7 @@ public final class ClassMetadata {
 	public static class GlobalConfiguration {
 		public static boolean ideFriendlyPacking;
 		public static boolean enabledPooledWeaving;
+		public static boolean enabledPackedWeaving;
 		public static boolean optimizeEntitySystems;
 	}
 }